Only log an out-of-memory warning after Gc fails
Before, we printed e.g. "memalign(4096, 4096) failed" if we couldn't
allocate a page, even if the only reason was that we needed to GC first.

Note that if we're also out of OCaml heap space then we might not be
able to print the warning at all. However, we still raise Out_of_memory
in that case.

See: https://github.com/mirage/mirage-tcpip/issues/33